+++ R/available-packages.R
@@ -358,12 +358,16 @@ renv_available_packages_latest <- function(package,
 {
   methods <- list(
     renv_available_packages_latest_repos,
-    renv_available_packages_latest_mran
+    if (renv_mran_enabled())
+      renv_available_packages_latest_mran
   )
 
   errors <- stack()
 
   entries <- lapply(methods, function(method) {
+
+    if (is.null(method))
+      return(NULL)
 
     entry <- catch(method(package, type, repos))
     if (inherits(entry, "error")) {
@@ -552,7 +556,7 @@ renv_available_packages_cellar <- function(type, project = NULL) {
   project <- renv_project_resolve(project)
   roots <- renv_cellar_roots(project = project)
 
-  # find all files used in the locals folder
+  # look for packages
   all <- list.files(
     path         = roots,
     all.files    = TRUE,
@@ -565,36 +569,30 @@ renv_available_packages_cellar <- function(type, project = NULL) {
   ext <- renv_package_ext(type = type)
   keep <- all[fileext(all) %in% ext]
 
-  # read the DESCRIPTION files within the archive
-  descs <- lapply(keep, function(path) {
+  # construct records for each cellar entry
+  records <- lapply(keep, function(path) {
 
-    # read the DESCRIPTION
-    desc <- renv_description_read(path)
+    # infer package name, version from tarball name
+    base <- basename(keep)
+    idx <- regexpr("_", base, fixed = TRUE)
+    package <- substring(base, 1L, idx - 1L)
+    version <- substring(base, idx + 1L, nchar(base) - nchar(ext))
 
     # set the Repository field
     prefix <- if (renv_platform_windows()) "file:///" else "file://"
-    uri <- paste0(prefix, dirname(path))
-    desc[["Repository"]] <- uri
+    repository <- paste0(prefix, dirname(path))
 
-    # return it
-    desc
-
-  })
-
-  # extract DESCRIPTION fields of interest
-  fields <- c("Package", "Version", "OS_type", "NeedsCompilation", "Repository")
-  records <- map(descs, function(desc) {
-
-    # ensure missing fields are set as NA
-    missing <- setdiff(fields, names(desc))
-    desc[missing] <- NA
-
-    # return record with requested fields
-    desc[fields]
+    # build record
+    list(
+      Package = package,
+      Version = version,
+      OS_type = NA,
+      NeedsCompilation = NA,
+      Repository = repository
+    )
 
   })
 
-  # bind into data.frame for lookup
   bind(records)
 
 }